requirements.txt
asgiref==3.8.1
boto3==1.35.14
botocore==1.35.14
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.8
jmespath==1.0.1
ntp==1.0.0a1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
vercel==0.2.1
whitenoise==6.7.0
